--- src/main/java/com/mineresetlite/InvalidCommandArgumentsException.java
@@ -1,8 +1,0 @@
-package com.koletar.jj.mineresetlite;
-
-/**
- * Thrown when a command's arguments were unsatisfactory.
- * @author jjkoletar
- */
-public class InvalidCommandArgumentsException extends Exception {
-}
